pyproject.toml
[project]
name = "sonic-o1-agent"
version = "0.1.0"
description = "Agentic audio-video processing with Qwen3-Omni"
readme = "README.md"
authors = [
    {name = "Ahmed Y. Radwan", email = "ahmed.radwan@vectorinstitute.ai"}
]
license = "Apache-2.0"
requires-python = ">=3.10"

dependencies = [
    # Core ML/DL frameworks
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "torchvision>=0.15.0",
    "transformers>=4.35.0",
    "accelerate>=0.20.0",

    # vLLM for efficient inference (required for Qwen3-Omni)
    "vllm>=0.5.0",
    "langgraph>=0.2.0",

    # Qwen3-Omni specific
    "qwen-omni-utils>=0.0.8",

    # Audio/Video processing
    "av>=10.0.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "opencv-python>=4.8.0",
    "ffmpeg-python>=0.2.0",
    "imageio>=2.0.0",

    # Data handling
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",

    # ML utilities
    "einops>=0.8.0",
    "timm>=0.9.0",
    "safetensors>=0.4.0",

    # Configuration and logging
    "pyyaml>=6.0.0",
    "python-dotenv>=1.0.0",

    # Progress and utilities
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
# Development tools
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.0.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "pre-commit>=3.0.0",
]

# Documentation
docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.0.0",
    "mkdocstrings[python]>=0.24.0",
]

# Full installation
all = [
    "sonic-o1-agent[dev,docs]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sonic_o1_agent"]

[tool.pytest.ini_options]
testpaths = ["tests/unit", "tests/integration"]
python_files = ["test_*.py"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 88
include = ["*.py", "pyproject.toml"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "N"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/sonic_o1_agent"]
omit = ["tests/*", "*__init__.py"]
